feat: seed-data.js writes 10M records to MongoDB
    - Batch size: 50000
    - Number of batches: 200
    - Total time: 20 minutes (laptop)
